+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog for NeoFS Node
 
 ### Fixed
 - Missing garbage, graveyard and to-move buckets in object metabase status (#3485)
+- Incorrect garbage bucket setting in metabase on tombstone writing (#3484)
 
 ### Changed